Package: Add 'start' script
Running the 'dev' script is problematic on Heroku as the 'watch' script prevents the process from completing. We only need to start a server and not watch for changes.
Therefore, create a version of the 'serve' script that does not watch files and just starts a server. The 'start' script is automatically run on Heroku and was not previously taken.