Add back Stylelint package and file
Add back the previously removed Stylelint package and file. Does not seem to be strictly necessary as the build completes without and there seems to be no issues with the webpage rendering (it's only a linter after all). However, it is used in a build step checking the SCSS files, so should be kept after all.